Move full filter menu to be section-specific
