+++ bmh_lims/frontend/src/components/index.js
@@ -9,4 +9,20 @@ export { MultilineText } from './MultilineText/'
 export { Checkbox } from './Checkbox/'
 export { default as CircularButtonBar } from './CircularButtonBar/'
 export { default as DropdownMenu, UnderlineDropdown } from './DropdownMenu/'
-export { default as FilterMenu } from './FilterMenu/'
+export {
+    FilterHeader,
+    FilterRow,
+    FilterContainer,
+    FreeTextFilter,
+    FreeTextFilterContainer,
+    FilterMenuContainer,
+    StyledDatePicker,
+    StyledUpArrow,
+    StyledDownArrow,
+    DateRangeContainer,
+    DateRangeFilterContainer,
+    StyledClearButton,
+    Filter,
+    DropdownFilter,
+    DateRangeFilter
+} from './FilterMenu'
