+++ bmh_lims/frontend/src/pages/Workflows/sections/AssignSection/components/index.js
@@ -1,0 +1,1 @@
+export { default as FilterMenu } from './FilterMenu'
